pyproject.toml
[project]
name = "mixdiscs"
version = "0.1.0"
description = "Playlists but nerdier!"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jinja2>=3.1.6",
    "pyyaml>=6.0.2",
    "spotipy>=2.25.0",
    "questionary>=2.0.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "ipdb>=0.13.13",
    "ipython>=8.32.0",
    "mypy>=1.15.0",
    "pylint>=3.3.4",
    "types-pyyaml>=6.0.12.20241230",
]
test = [
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "freezegun>=1.5.0",
]
pyyaml-build = ["cython<3", "setuptools"]

[tool.uv]
package = true
no-build-isolation-package = ["pyyaml"]

[tool.setuptools]
packages = ["mixdiscer", "mixdiscer.cli", "mixdiscer.music_service", "mixdiscer.output"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov-report=term-missing",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.coverage.run]
source = ["mixdiscer"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/.venv/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[project.scripts]
create-playlist = "mixdiscer.create_playlist:main"

